requirements.txt
streamlit
torch
torchvision
efficientnet_pytorch
Pillow
